Eliminated the last remnants of "command" (hopefully)
