+++ game/application/src/main/kotlin/de/gleex/pltcmd/game/application/examples/conversation/conversationExample.kt
@@ -1,11 +1,11 @@
 package de.gleex.pltcmd.game.application.examples.conversation
 
 import de.gleex.pltcmd.game.engine.Game
-import de.gleex.pltcmd.game.engine.commands.ConversationMessage
 import de.gleex.pltcmd.game.engine.entities.EntityFactory
 import de.gleex.pltcmd.game.engine.entities.types.ElementEntity
 import de.gleex.pltcmd.game.engine.entities.types.callsign
 import de.gleex.pltcmd.game.engine.entities.types.inConversationWith
+import de.gleex.pltcmd.game.engine.messages.ConversationMessage
 import de.gleex.pltcmd.game.options.GameOptions
 import de.gleex.pltcmd.game.options.UiOptions
 import de.gleex.pltcmd.game.ticks.Ticker
