+++ game/engine/src/main/kotlin/de/gleex/pltcmd/game/engine/attributes/goals/Goal.kt
@@ -10,7 +10,7 @@ import org.hexworks.cobalt.logging.api.LoggerFactory
  * An element may have a _goal_ that is an abstraction layer on top of the basic capabilities of "a bunch of soldiers".
  *
  * It gets [step]ped until it [isFinished]. Goals may contain sub-goals so each goal can represent any
- * level of abstraction until it breaks down to a [Command] that needs to be executed by the entity at the lowest level.
+ * level of abstraction until it breaks down to a [Message] that needs to be executed by the entity at the lowest level.
  *
  * Sub-goals work like an ordered queue. The head of the queue  is the currently active sub-goal and is removed as soon
  * as it is finished. Then the next queued goal with [appendSubGoals] gets active (FIFO, first-in-first-out).
@@ -35,9 +35,9 @@ abstract class Goal(vararg subGoals: Goal) {
     }
 
     /**
-     * Advances this goal by one step. It may return a command that needs to be executed by the element.
+     * Advances this goal by one step. It may return a message that needs to be executed by the element.
      *
-     * It should not call execute/sendCommand directly. At most this method might set attributes.
+     * It should not call receive/sendMessage directly. At most this method might set attributes.
      *
      * The default implementation simply advances the sub-goals by calling [stepSubGoals].
      */
