+++ game/engine/src/main/kotlin/de/gleex/pltcmd/game/engine/messages/MoveTo.kt
@@ -1,14 +1,14 @@
-package de.gleex.pltcmd.game.engine.commands
+package de.gleex.pltcmd.game.engine.messages
 
 import de.gleex.pltcmd.game.engine.GameContext
 import de.gleex.pltcmd.game.engine.entities.types.MovableEntity
 import de.gleex.pltcmd.model.world.coordinate.Coordinate
 import org.hexworks.amethyst.api.Message
 
-/** Message that is fired when the position of an entity changes. It contains the old and new position. */
-data class UpdatePosition(
-        val oldPosition: Coordinate,
-        val newPosition: Coordinate,
+
+/** Message to provide entities a destination **/
+data class MoveTo(
+        val destination: Coordinate,
         override val context: GameContext,
         override val source: MovableEntity
 ) : Message<GameContext>
