+++ game/engine/src/main/kotlin/de/gleex/pltcmd/game/engine/systems/behaviours/IntentPursuing.kt
@@ -16,11 +16,11 @@ object IntentPursuing : BaseBehavior<GameContext>(CommandersIntent::class) {
     override suspend fun update(entity: AnyGameEntity, context: GameContext): Boolean {
         return if (entity.type == ElementType) {
             val element = entity as ElementEntity
-            val commandToExecute = element.commandersIntent.proceed(element, context)
-            if (commandToExecute.isPresent) {
-                element.receiveMessage(commandToExecute.get())
+            val messageToExecute = element.commandersIntent.proceed(element, context)
+            if (messageToExecute.isPresent) {
+                element.receiveMessage(messageToExecute.get())
             }
-            commandToExecute.isPresent
+            messageToExecute.isPresent
         } else {
             false
         }
