+++ game/engine/src/main/kotlin/de/gleex/pltcmd/game/engine/systems/behaviours/LookingAround.kt
@@ -3,9 +3,9 @@ package de.gleex.pltcmd.game.engine.systems.behaviours
 import de.gleex.pltcmd.game.engine.GameContext
 import de.gleex.pltcmd.game.engine.attributes.PositionAttribute
 import de.gleex.pltcmd.game.engine.attributes.VisionAttribute
-import de.gleex.pltcmd.game.engine.commands.DetectEntities
 import de.gleex.pltcmd.game.engine.entities.EntitySet
 import de.gleex.pltcmd.game.engine.entities.types.*
+import de.gleex.pltcmd.game.engine.messages.DetectEntities
 import de.gleex.pltcmd.model.elements.Affiliation
 import de.gleex.pltcmd.model.signals.vision.builder.visionAt
 import org.hexworks.amethyst.api.base.BaseBehavior
@@ -15,7 +15,7 @@ import org.hexworks.cobalt.logging.api.LoggerFactory
 
 /**
  * Behavior of an entity that updates the [VisionAttribute] each tick if needed. It also
- * sends a [DetectEntities] command if anything is present in the vision.
+ * sends a [DetectEntities] message if anything is present in the vision.
  **/
 object LookingAround :
         BaseBehavior<GameContext>(
