+++ game/engine/src/main/kotlin/de/gleex/pltcmd/game/engine/systems/facets/ConversationSender.kt
@@ -2,8 +2,8 @@ package de.gleex.pltcmd.game.engine.systems.facets
 
 import de.gleex.pltcmd.game.engine.GameContext
 import de.gleex.pltcmd.game.engine.attributes.RadioAttribute
-import de.gleex.pltcmd.game.engine.commands.ConversationMessage
 import de.gleex.pltcmd.game.engine.entities.types.startConversation
+import de.gleex.pltcmd.game.engine.messages.ConversationMessage
 import org.hexworks.amethyst.api.Consumed
 import org.hexworks.amethyst.api.Response
 import org.hexworks.amethyst.api.base.BaseFacet
